+++ pkg/host/config.go
@@ -22,9 +22,6 @@ type SonrHost interface {
 	// Address returns the account address of the underlying wallet for the host
 	Address() string
 
-	// AddrInfo returns the Libp2p hosts peer.AddrInfo as a Sonr Common AddrInfo
-	AddrInfo(protocols ...protocol.ID) ct.AddrInfo
-
 	// Config returns the configuration of the node
 	Config() *config.Config
