+++ proto/service/v1/discover.proto
@@ -10,14 +10,14 @@ option java_package = "io.sonr.types.service";
 option java_multiple_files = true;
 
 // LobbyMessage is message passed from Peer in Lobby
-message PubSubLobbyMessage {
+message LobbyMessage {
   common.v1.Peer from = 1;     // Users Peer Data
   bytes body = 2;    // Payload is the message data
   bytes signature = 3;             // Signature of the Attached Data
   int64 created_at = 4; // Timestamp
 }
 
-message DiscoverEvent {
+message RefreshEvent {
   string topic_name = 1;
   repeated common.v1.Peer peers = 2;
   int64 received_at = 3;
